# tools/api-docs/public/openapi.yaml
openapi: 3.0.3
info:
  title: Business Data Streaming Pipeline — HTTP APIs
  description: |
    Tài liệu tương tác cho demo k3s. **Chọn đúng Server** (góc trên) trước Try it out.

    Request đi qua **Vite proxy** (cùng origin `:5190`, tránh CORS). Chọn Server:

    | Server | Backend thật |
    |--------|----------------|
    | `/proxy/dashboard` | dashboard-api :32000 |
    | `/proxy/tracking` | tracking-api :31000 |
    | `/proxy/commerce` | commerce-backend :30330 |

    Backend trên WSL IP: tạo `tools/api-docs/.env` với `VITE_PROXY_DASHBOARD=http://<WSL_IP>:32000` (và tương tự tracking/commerce), restart `npm run dev`.

    Tham số kỳ: `minutes` (rolling) hoặc `date=YYYY-MM-DD` (UTC, ưu tiên hơn minutes).
  version: 1.0.0
servers:
  - url: /proxy/dashboard
    description: dashboard-api — qua Vite proxy → :32000
  - url: /proxy/tracking
    description: tracking-api — qua Vite proxy → :31000
  - url: /proxy/commerce
    description: commerce-backend — qua Vite proxy → :30330
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard login — POST /api/auth/login
    ingestKey:
      type: http
      scheme: bearer
      description: TRACKING_INGEST_API_KEY (Lap2 adapter)
...
  schemas:
...
    TrackingEvent:
      type: object
      required:
        - event_id
        - event_time
        - event_type
        - event_source
        - event_category
        - session_id
      properties:
        event_id: { type: string }
        event_time: { type: string, format: date-time }
        event_type: { type: string }
        event_source: { type: string }
        event_category:
          type: string
          enum: [behavior, commerce]
        session_id: { type: string }
        anonymous_id: { type: string }
        user_id: { type: string }
        page_url: { type: string }
        product_id: { type: string }
        metadata: { type: object, additionalProperties: true }
    TrackBatch:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/TrackingEvent" }
...
paths:
...
  /track:
    post:
      tags: [tracking]
      servers: [{ url: /proxy/tracking }]
      summary: Ingest một tracking event (SDK)
      description: Không API key — chỉ validate schema
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TrackingEvent" }
      responses:
        "202": { description: accepted }
        "400": { description: validation_failed }
  /track/batch:
    post:
      tags: [tracking]
      servers: [{ url: /proxy/tracking }]
      summary: Ingest batch events
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TrackBatch" }
      responses:
        "202": { description: accepted }
  /api/ingest/business-events:
    post:
      tags: [tracking]
      servers: [{ url: /proxy/tracking }]
      summary: Business event ingest (adapter Lap2)
      security: [{ ingestKey: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: "Canonical business event hoặc batch object events[]"
      responses:
        "202": { description: accepted }
        "401": { description: unauthorized }
        "503": { description: business_ingest_disabled }
  /api/ingest/business-events/batch:
    post:
      tags: [tracking]
      servers: [{ url: /proxy/tracking }]
      summary: Business events batch
      security: [{ ingestKey: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                events:
                  type: array
                  items: { type: object }
      responses:
        "202": { description: accepted }
...